requirements.txt
torch
torchvision
transformers
albumentations
scikit-learn
sam-model  # placeholder for SAM library
Pillow
tqdm 